* NEW options page to toggle debug settings, plugin dependencies, dashboard widgets, editor metaboxes, and more * NEW daily cron task, use action hook 'soma_daily_event' * NEW csv export class, included in trunk but not ready for prime time * NEW plugin checks for minimum wordpress version, aborts otherwise... * sexy settings link right on the plugins page! * improved the example php config templates
git-svn-id: https://plugins.svn.wordpress.org/somatic-framework/trunk@538041 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ inc/somaFunctions.php
@@ -896,7 +896,9 @@ class somaFunctions extends somaticFramework {
 	//** checks for activation of plugins that somaFramework is dependent on ------------------------------------------------------//
 	function check_plugin_dependency() {
 		// require scribu's p2p plugin
-		if ( !function_exists('p2p_register_connection_type') ) {
+		$opt = get_option('somatic_framework_options');
+		
+		if ( $opt['p2p'] && !function_exists('p2p_register_connection_type') ) {
 			add_action( 'admin_notices', create_function('', "
 				echo '<div id=\"message\" class=\"error\" style=\"font-weight: bold\"><p>PLUGIN REQUIRED: \"Posts 2 Posts\" - please <a href=\"http://scribu.net/wordpress/posts-to-posts\" target=\"_blank\">download</a> and/or activate!</p></div>';
 			"));
@@ -937,7 +939,7 @@ class somaFunctions extends somaticFramework {
 			add_action('admin_notices', 'wp_version_update' );
 			function wp_version_update() {
 				echo '<div id="message" class="error" style="font-weight: bold">';
-				echo '<p>WORDPRESS 3.2 MINIMUM REQUIRED - please update WP or de-activate this plugin!</p>';
+				echo '<p>WORDPRESS 3.3 MINIMUM REQUIRED - please update WP or de-activate this plugin!</p>';
 				echo '</div>';
 			}
 		}
